chore: remove unused and relocate emoji picker
