+++ apps/web/app/(base)/get-started/EmojiPicker.tsx
